Added --with-d-bus option to Qt5 dependency when we build cpp-ethereum --with-gui.
Also reordered the options to be BEFORE the dependencies.   My suspicion is that the conditional dependency for Qt5 never actually worked, but that it had never been tested in a clean environment.   Looks like build.with? is dependent on those options having been declared, based on local testing of this Formula.