Update instructions for prioritizing latest keg
`tap-pin` was deprecated in https://brew.sh/2019/04/04/homebrew-2.1.0/
and is no longer available. The official recommentation is to either
install mentioning the tap explicitly, or to rename the formula in one
place to no longer have the naming conflict.

Also fixed the link to the `homebrew-core` keg. They made subdirectories
for the first letter of the formula at some point.